requirements.txt
asttokens==2.0.5
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
decorator==5.1.1
executing==0.8.3
Flask==2.1.2
fonttools==4.33.3
idna==3.3
importlib-metadata==4.11.4
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.23.0
opencv-python==3.4.17.63
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.0
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
stack-data==0.3.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.0
